+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support `global-state` expressions in `sky.*` properties ([#7967](https://github.com/maplibre/maplibre-gl-js/pull/7967)) (by [@CommanderStorm](https://github.com/CommanderStorm))
 - Support `global-state` expressions in `light.*` properties ([#7966](https://github.com/maplibre/maplibre-gl-js/pull/7966)) (by [@CommanderStorm](https://github.com/CommanderStorm))
 - Add `MapOptions.rotateSpeed` and `MapOptions.pitchSpeed`, the degrees the bearing/pitch change per pixel dragged ([#7949](https://github.com/maplibre/maplibre-gl-js/pull/7949)) (by [@clement-igonet](https://github.com/clement-igonet))
+- Add opt-in quaternion (versor) globe panning via `DragPanOptions.fixedBearing`, for smooth dragging near and across the poles ([#8017](https://github.com/maplibre/maplibre-gl-js/pull/8017)) (by [@jcolot](https://github.com/jcolot))
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes